backend/requirements.txt
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==21.2.0
blinker==1.8.0
cachetools==5.3.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
comm==0.2.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==3.2.0
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
more-itertools==8.10.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openai==1.23.6
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pg8000==1.30.5
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.2
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
scramp==1.4.4
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
systemd-python==234
texttable==1.6.4
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
websocket-client==1.2.3
Werkzeug==3.0.2
zipp==1.0.0
